[ANDROAPP-3882] Refactor FieldUiModel.kt to form library
