+++ app/src/main/java/org/dhis2/data/forms/dataentry/DataEntryAdapter.kt
@@ -7,11 +7,12 @@ import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.ListAdapter
 import java.util.ArrayList
 import java.util.LinkedHashMap
-import org.dhis2.data.forms.dataentry.fields.FieldUiModel
 import org.dhis2.data.forms.dataentry.fields.FieldViewModel
 import org.dhis2.data.forms.dataentry.fields.FormViewHolder
 import org.dhis2.data.forms.dataentry.fields.FormViewHolder.FieldItemCallback
 import org.dhis2.data.forms.dataentry.fields.section.SectionViewModel
+import org.dhis2.form.data.FieldUiModel
+import org.dhis2.form.ui.DataEntryDiff
 
 class DataEntryAdapter :
     ListAdapter<FieldUiModel, FormViewHolder>(DataEntryDiff()),
