+++ app/src/main/java/org/dhis2/data/forms/dataentry/fields/edittext/EditTextViewModel.java
@@ -10,15 +10,17 @@ import com.google.auto.value.AutoValue;
 import org.dhis2.R;
 import org.dhis2.data.forms.dataentry.DataEntryViewHolderTypes;
 import org.dhis2.data.forms.dataentry.fields.ActionType;
-import org.dhis2.data.forms.dataentry.fields.FieldUiModel;
 import org.dhis2.data.forms.dataentry.fields.FieldViewModel;
-import org.dhis2.data.forms.dataentry.fields.KeyboardActionType;
+
+import org.dhis2.form.data.KeyboardActionType;
+
 import org.dhis2.data.forms.dataentry.fields.LegendValue;
 import org.dhis2.data.forms.dataentry.fields.RowAction;
 import org.hisp.dhis.android.core.common.ObjectStyle;
 import org.hisp.dhis.android.core.common.ValueType;
 import org.hisp.dhis.android.core.common.ValueTypeDeviceRendering;
 
+import org.dhis2.form.data.FieldUiModel;
 import io.reactivex.processors.FlowableProcessor;
 
 /**
