+++ app/src/main/java/org/dhis2/data/forms/dataentry/fields/radiobutton/RadioButtonViewModel.java
@@ -8,7 +8,6 @@ import com.google.auto.value.AutoValue;
 import org.dhis2.R;
 import org.dhis2.data.forms.dataentry.DataEntryViewHolderTypes;
 import org.dhis2.data.forms.dataentry.fields.ActionType;
-import org.dhis2.data.forms.dataentry.fields.FieldUiModel;
 import org.dhis2.data.forms.dataentry.fields.FieldViewModel;
 import org.dhis2.data.forms.dataentry.fields.RowAction;
 import org.hisp.dhis.android.core.common.ObjectStyle;
@@ -18,6 +17,7 @@ import org.hisp.dhis.android.core.common.ValueTypeRenderingType;
 import java.util.Locale;
 import java.util.Objects;
 
+import org.dhis2.form.data.FieldUiModel;
 import io.reactivex.processors.FlowableProcessor;
 
 @AutoValue
