+++ form/src/main/java/org/dhis2/form/data/FieldUiModel.kt
@@ -1,4 +1,4 @@
-package org.dhis2.data.forms.dataentry.fields
+package org.dhis2.form.data
 
 interface FieldUiModel {
