+++ form/src/main/java/org/dhis2/form/ui/DataEntryDiff.kt
@@ -1,7 +1,7 @@
-package org.dhis2.data.forms.dataentry
+package org.dhis2.form.ui
 
 import androidx.recyclerview.widget.DiffUtil
-import org.dhis2.data.forms.dataentry.fields.FieldUiModel
+import org.dhis2.form.data.FieldUiModel
 
 class DataEntryDiff : DiffUtil.ItemCallback<FieldUiModel>() {
     override fun areItemsTheSame(oldItem: FieldUiModel, newItem: FieldUiModel): Boolean =
